New function sphrFaceNormals
The original albedo function calculates relevant face normal vectors with the functions idx2radian() and sphr2cart() every time the function is called in a for loop. Making a map of all sphere face normal vectors indexed corresponding to the Earth reactivity map is what this function dose. This is needed for vectorising the albedo calculation and to avoid finding the same normal vectors again and again.