New Function sphrFaceAreas()
Dan D. V. Bhanderi cellarea() function was used repeatedly to find the same cellarea/facearea each time the albedo function was called. This function uses the same algorithm in a vectorised manner to make a column(all unique areas[m^2])  corresponding to the columns of the reflectivity map. Facilitating vectorisation of the the albedo function