BugFix: Incorrect FOV box estimate fixed
The box estimate of FOV did not grow fast enough as the satellite approached a pole. Equation for expansion of the box in equatorial/east-west/longitude direction is now calculated at the edge of the latitude box estimate. Now there is no difference between a test of all Earth faces and a test limited to the box estimate. How tight/efficient this adjusted estimate is has not been tested.